requirements.txt
pyside6
pandas
openpyxl
pyinstaller
requests
packaging
beautifulsoup4
google-generativeai
lxml
Pillow
